prosody_dynamic() bug fixed in prosody.py
tempvec.append(dur) moved from line  403 to line 414 after tempvec.extend(z.coeffs).
In this way the duration of the audio segment is actually the last of the 13 features. Otherwise it would have been the first one under the name of 'f0coef0'. This would shift all the features by one position and get them all with the wrong name.